+++ changelog.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.2.0 -2020-01-21
+## Added 
+- _POST_FLAT_FILE_FULFILLMENT_DATA_ feed integration
+- MWSOrderFulfilment Object
+
+## Changed
+- Reverted MWS->product() back to hard coded array formate as private variables are pe-fixed with namespace and returned as array elements. This causes issues..
+- Added a secodary header row as Amazon appears to want this..
+
 ## 1.1.0 - 2019-10-08
 ### Added
 - Ability to change a config variable at run-time.